Add FF-HEDM single-grain optimization demo with clean and noisy modes.
Uses resolution-based sigma weights derived from detector precision (sigma_2th=0.5px/Lsd, sigma_eta=0.5/R_ring, sigma_ome=0.25omega_step) for physically motivated coordinate balancing. Three-phase L-BFGS (orientation/strain/joint) recovers orientation to 0.008 deg and lattice to 1.6e-3 A from 1.7 deg perturbation with realistic noise. 91 unit tests pass.

+++ fwd_sim/hedm_forward.py
@@ -398,8 +398,8 @@ class HEDMForwardModel(nn.Module):
         ], dim=-2)
 
         # G_cart = B @ hkls_int^T  =>  (..., M, 3)
-        # hkls_int is (M, 3), B is (..., 3, 3)
-        hkls_cart = torch.einsum("...ij,mj->...mi", B, self.hkls_int)
+        # hkls_int is (M, 3), B is (..., 3, 3) -- match dtype
+        hkls_cart = torch.einsum("...ij,mj->...mi", B, self.hkls_int.to(B.dtype))
 
         # d-spacing = 1 / |G_cart|
         g_norm = torch.norm(hkls_cart, dim=-1).clamp(min=self.epsilon)
